requirements.txt
appdirs==1.4.4
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
idna==3.2
mypy-extensions==0.4.3
pathspec==0.9.0
pkg_resources==0.0.0
psycopg2-binary==2.9.1
regex==2021.8.3
requests==2.26.0
tomli==1.2.1
urllib3==1.26.6
